requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
networkx>=3.1
matplotlib>=3.7.0
seaborn>=0.12.0
nltk>=3.8.1
vaderSentiment>=3.3.2
pyyaml>=6.0
kagglehub>=0.1.0
python-louvain>=0.16
pyarrow>=12.0.0
pytest>=7.4.0
pytest-mock>=3.10.0
pytest-cov>=4.1.0